requirements.txt
# Requirements for luzeiros

# Django
django>=2.1.5


# Error logging
sentry-sdk==0.7.11


# Models
django-recurrence


# Event Queueing, Web workers
gunicorn

# Databases
psycopg2-binary
djongo


# Cloud
cloudinary
boto3
django-storages


# Image Support
Pillow


# WYSIWYG Editor
django-ckeditor


# Authentication + Permissions
django-role-permissions
django-registration
django-otp
django-two-factor-auth
twilio
qrcode


# REST API Support
django-rest-swagger
djangorestframework>=3.7.7
djangorestframework-httpsignature
djangorestframework-jsonapi>=2.6.0
djangorestframework_simplejwt
django-filter>=2.0.0
django-guardian
django-oauth-plus
django-oauth2-provider
oauth2


# Server
django-debug-toolbar

# Chats
channels
